avcodec/vaapi_encode: Use refstruct pool API, stop abusing AVBuffer API
Up until now, the VAAPI encoder uses fake data with the
AVBuffer-API: The data pointer does not point to real memory,
but is instead just a VABufferID converted to a pointer.
This has probably been copied from the VAAPI-hwcontext-API
(which presumably does it to avoid allocations).

This commit changes this without causing additional allocations
by switching to the refstruct-pool API.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>