avcodec/mpegpicture: Make ff_find_unused_picture return Picture*
Currently it returns an index into an array of Pictures.
Yet all callers only use this index for exactly one purpose:
To get the pointer to the actual picture. So modify
ff_find_unused_picture() to return said pointer directly.

This is in preparation for future commits where there won't
be an array of Pictures any more.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>