avcodec/threadprogress: Add new API for frame-threaded progress
The API is very similar by the ThreadFrame API, with the exception
that it no longer has an included AVFrame. Instead one can wait
on anything via a ThreadProgress. One just has to ensure that
the lifetime of the object containing the ThreadProgress is long
enough (the corresponding problem for ThreadFrames is solved
by allocating the progress and giving each thread a reference
to it). This will typically be solved by putting a ThreadProgress
in a refcounted structure that is shared between threads.
It will be put to the test in the VP9 decoder.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>