avcodec/vp9: Switch to ProgressFrames
This already fixes a race in the vp9-encparams test. In this test,
side data is added to the current frame after having been decoded
(and therefore after ff_thread_finish_setup() has been called).
Yet the update_thread_context callback called ff_thread_ref_frame()
and therefore av_frame_ref() with this frame as source frame and
the ensuing read was unsynchronised with adding the side data,
i.e. there was a data race.

By switching to the ProgressFrame API the implicit av_frame_ref()
except if this frame is later reused by a show-existing-frame.
The vp9-encparams test does not cover this, so this commit
already fixes all the races in this test.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>