avcodec/metasound_data: Split file into metasound-only and shared stuff
Currently, lavc/metasound_data.c contains two types of data:
TwinVQModeTabs that are only used by the metasound decoder
and some arrays of floats that are used both by MetaSound
and TwinVQ. This commit changes this: The MetaSound-only
data is moved into metasound_data.h which is now directly
included into metasound.c (the TwinVQModeTabs are made static).
The rest is moved into a new file, metasound_twinvq_data.h
which is directly included into twinvq.c (the common file
of the MetaSound and TwinVQ decoders).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>