avcodec/vc1: Don't check for AVCodecContext.codec
This check has been added in c617bed34f39a122ab1f89581ddce9cc63885383,
merging ee769c6a7c1d4ec6560f5e5a6f457b770b10fb33 to fix
a possible segfault if AVCodecContext.codec is not set
as it may be during parsing. While this fixes the segfault,
it has the unfortunate side effect that it makes the output
of the parser dependent on whether a decoder is set (and
ultimately available). The fix later applied in
5d2be71b9ecf2a88752666a2c4039f4d98419d35 does not have this
downside and makes checking AVCodecContext.codec superfluous.
So remove this check.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>