avcodec/get_bits: Avoid reloading when reading VLC when possible
When reading VLCs with depth > 1, the current GET_VLC macro
updates the cache (i.e. reloads from the buffer) at every
depth. Yet this is usually wasteful, as most VLCs are shorter
than the initially available 25 bits and could be read without
reloading (and even for most of those where this is not so,
reloading at the second stage is unnecessary).

This could be fixed by giving GET_VLC precise information
about the length of the maximum VLC element, yet this
information is not readily available (it would require
a replacement of get_vlc2()).

But we nevertheless have an upper bound on the number of bits
to be read (maximally) at every stage: It is stage * num of bits
to be read in the first stage. This commit uses this information,
but only if it is a compile-time constant in order to avoid
runtime checks. Doing so saved 7248 B from .text here (GCC 11.3;
it were 3280 B with Clang 14 and 1296 B with Clang 15 (all -O3)).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>