avcodec/h264dec: Check early whether film grain is supported, fix race
Applying film grain happens after ff_thread_finish_setup(),
so the parameters synced in ff_h264_update_thread_context()
must not be modified. But this is exactly what happens if
applying film grain fails. (The likely result is that in
case of frame threading an uninitialized frame is output.)

Failure of applying film grain can have several reasons:
The first is that it is just not supported (we only support
a specific model_id and only YUV420p). The second is that
no film grain side data is added to the AVFrame. This seems
possible even with legal input, namely if the second field
overrides (i.e. present equal to zero) an earlier film grain SEI.

Given that it is actually very easy to know in advance whether
ff_h274_apply_film_grain() supports a given set of parameters,
one can check for this before ff_thread_finish_setup()
and avoid allocating an unused buffer lateron.

The second failure condition has been addressed by keeping
track of the film grain state in a more granular fashion,
namely by distinguishing the cases of an allocated film
grain frame buffer and the existence of the side data.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>