avcodec/h264dec: Share fields between H264Pictures/threads directly
To do this, add a common structure (H264SharedPic) that is
propagated across ff_h264_ref_frame() and ff_h264_replace_frame().
The structure is RefStruct-enabled and owns the references
contained in it. This meant that the base-pointers could
be moved to it without replacement; the non-base pointers stay
as non-references.

It also means that one can combine several ff_refstruct_replace()
into one in ff_h264_(ref|replace)_frame().

hwaccel_picture_private has also been moved to it.

The H264SharedPictures are part of a RefStruct pool; this would
not be possible with an AVBufferPool due to the latter's lack
of a reset callback.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>