avcodec/rkmppdec: Fix double-free on error
After having created the AVBuffer that is put into
frame->buf[0], ownership of several objects
(namely an AVDRMFrameDescriptor and an MppFrame
as well framecontextref and decoder_ref (owned by framecontextref))
passes to the AVBuffer and therefore to the frame.
Yet it has nevertheless been freed manually on error
afterwards, which would lead to a double-free as soon
as the AVFrame is unreferenced.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>