avcodec/mpegvideo_enc: Avoid allocs, creating new AVFrame references
The earlier code allocated some accessory buffers for frames
before it was certain that they would be used at all (some were not
in case rc_buffer_size is set); it also created a new reference
via av_frame_ref() even though it sometimes threw the old one
away a few lines later.

This commit fixes both of this: Allocating the accessory buffers
is now deferred to the latest point possible; this means that
the Picture containing the AVFrame is now basically blank besides
the AVFrame, so if we just move the reference instead of creating
a new one, we end up with a blank Picture that can be reused.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>