avcodec/mpegvideo_dec, mpeg12dec: Move MPEG1/2 code to mpeg12dec.c
Up until now, ff_mpv_frame_start() offsets the data of the current
picture and doubles the linesizes of all pictures if the current
picture is field-based so that data and linesize allow to address
the current field only.

This is done based upon the current picture_structure value.
Only two mpegvideo-based decoders ever set this field: mpeg1/2
and VC-1; but the latter only does it when using hardware-acceleration
in which case the offset applied to these values is irrelevant
as said pointers (and the linesizes) are not read. So only one
decoder actually wants this offset; therefore move the code
performing it to mpeg12dec.c.

Also only offset the first three pointers, as MPEG-1/2 does not
support alpha.

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>