avcodec/aacdec: Use same AACDecContext for fixed and float
Up until now, there was one AACDecContext for the fixed
and one for the floating point decoder. These differed
mostly in certain arrays which were int for the fixed-point
and float for the floating point decoder; there were also
differences in corresponding function pointers.

Yet in order to deduplicate the enormous amount of currently
duplicated code between the float and the fixed-point decoder,
one needs common contexts. Given that int and float have the
same size on all common systems, this commit replaces these
arrays by unions of int arrays and of float arrays. Accesses
to these arrays are redirected to the proper target by defines
(with the exception of ChannelCoupling.gain, which clashes with
SpectralBandReplication.gain).

Signed-off-by: Andreas Rheinhardt <andreas.rheinhardt@outlook.com>